Fix cli-kit tests
