+++ packages/cli-kit/src/private/node/session/exchange.ts
@@ -3,7 +3,7 @@ import {applicationId, clientId as getIdentityClientId} from './identity.js'
 import {CodeAuthResult} from './authorize.js'
 import {API} from '../../../network/api.js'
 import {identityFqdn} from '../../../public/node/environment/fqdn.js'
-import {shopifyFetch} from '../../../http.js'
+import {shopifyFetch} from '../../../public/node/http.js'
 import {err, ok, Result} from '../../../public/node/result.js'
 import {AbortError} from '../../../public/node/error.js'
 import {ExtendableError} from '../../../error.js'
