+++ packages/cli-kit/src/private/node/session/identity-token-validation.ts
@@ -1,6 +1,6 @@
 import {identityFqdn} from '../../../public/node/environment/fqdn.js'
 import {debug} from '../../../output.js'
-import {shopifyFetch} from '../../../http.js'
+import {shopifyFetch} from '../../../public/node/http.js'
 
 export async function validateIdentityToken(token: string) {
   try {
