+++ packages/cli-kit/src/public/node/api/admin.test.ts
@@ -2,7 +2,7 @@ import * as admin from './admin.js'
 import {AdminSession} from '../session.js'
 import {graphqlRequest} from '../../../private/node/api/graphql.js'
 import {buildHeaders} from '../../../private/node/api/headers.js'
-import * as rest from '../../../private/node/api/rest.js'
+import * as http from '../../../public/node/http.js'
 import {test, vi, expect, describe} from 'vitest'
 
 vi.mock('../../../private/node/api/graphql.js')
@@ -65,7 +65,7 @@ describe('admin-rest-api', () => {
     const status = 200
     const headers = {'x-request-id': 123}
 
-    vi.spyOn(rest, 'fetch').mockResolvedValue({
+    vi.spyOn(http, 'fetch').mockResolvedValue({
       json,
       status,
       headers: {raw: () => headers},
@@ -86,8 +86,8 @@ describe('admin-rest-api', () => {
     const status = 200
     const headers = {'X-Shopify-Access-Token': `Bearer ${token}`, 'Content-Type': 'application/json'}
 
-    vi.mocked(buildHeaders).mockResolvedValue(headers)
-    const spyFetch = vi.spyOn(rest, 'fetch').mockResolvedValue({
+    vi.mocked(buildHeaders).mockReturnValue(headers)
+    const spyFetch = vi.spyOn(http, 'fetch').mockResolvedValue({
       json,
       status,
       headers: {raw: () => ({})},
@@ -113,8 +113,8 @@ describe('admin-rest-api', () => {
     const status = 200
     const headers = {'X-Shopify-Access-Token': `Bearer ${token}`, 'Content-Type': 'application/json'}
 
-    vi.mocked(buildHeaders).mockResolvedValue(headers)
-    const spyFetch = vi.spyOn(rest, 'fetch').mockResolvedValue({
+    vi.mocked(buildHeaders).mockReturnValue(headers)
+    const spyFetch = vi.spyOn(http, 'fetch').mockResolvedValue({
       json: () => Promise.resolve({result: true}),
       status,
       headers: {raw: () => ({})},
