+++ packages/cli-kit/src/public/node/api/oxygen.test.ts
@@ -1,5 +1,5 @@
 import * as oxygenApi from './oxygen.js'
-import {shopifyFetch, formData} from '../../../http.js'
+import {shopifyFetch, formData} from '../http.js'
 import {buildHeaders} from '../../../private/node/api/headers.js'
 import {graphqlRequest} from '../../../private/node/api/graphql.js'
 import {test, vi, describe, beforeEach, expect} from 'vitest'
@@ -7,7 +7,7 @@ import {Response} from 'node-fetch'
 
 vi.mock('../../../private/node/api/graphql.js')
 vi.mock('../../../private/node/api/headers.js')
-vi.mock('../../../http.js')
+vi.mock('../http.js')
 
 const mockedResult = 'OK'
 const mockedToken = 'token'
