+++ packages/cli-kit/src/public/node/http.ts
@@ -1,11 +1,32 @@
-import {httpsAgent} from '../http.js'
-import {content, debug} from '../output.js'
-import {buildHeaders, sanitizedHeadersOutput} from '../private/node/api/headers.js'
-import nodeFetch from 'node-fetch'
-import {performance} from 'perf_hooks'
-import type {RequestInfo, RequestInit} from 'node-fetch'
+import {buildHeaders, httpsAgent, sanitizedHeadersOutput} from '../../private/node/api/headers.js'
+import {content, debug} from '../../output.js'
+import FormData from 'form-data'
+import nodeFetch, {RequestInfo, RequestInit} from 'node-fetch'
 
-type Response = ReturnType<typeof nodeFetch>
+export {
+  createApp,
+  createRouter,
+  IncomingMessage,
+  ServerResponse,
+  CompatibilityEvent,
+  createError,
+  send,
+  sendError,
+  sendRedirect,
+  H3Error,
+} from 'h3'
+
+/**
+ * Create a new FormData object.
+ *
+ * @returns A FormData object.
+ */
+export function formData(): FormData {
+  return new FormData()
+}
+
+export type Response = ReturnType<typeof nodeFetch>
+
 /**
  * An interface that abstracts way node-fetch. When Node has built-in
  * support for "fetch" in the standard library, we can drop the node-fetch
@@ -13,11 +34,12 @@ type Response = ReturnType<typeof nodeFetch>
  * Note that we are exposing types from "node-fetch". The reason being is that
  * they are consistent with the Web API so if we drop node-fetch in the future
  * it won't require changes from the callers.
+ *
  * @param url - This defines the resource that you wish to fetch.
- * @param init - An object containing any custom settings that you want to apply to the request
+ * @param init - An object containing any custom settings that you want to apply to the request.
  * @returns A promise that resolves with the response.
  */
-export default async function fetch(url: RequestInfo, init?: RequestInit): Response {
+export async function fetch(url: RequestInfo, init?: RequestInit): Response {
   const response = await nodeFetch(url, init)
   return response
 }
@@ -25,13 +47,17 @@ export default async function fetch(url: RequestInfo, init?: RequestInit): Respo
 /**
  * A fetch function to use with Shopify services. The function ensures the right
  * TLS configuragion is used based on the environment in which the service is running
- * (e.g. spin)
+ * (e.g. Spin).
+ *
+ * @param url - This defines the resource that you wish to fetch.
+ * @param init - An object containing any custom settings that you want to apply to the request.
+ * @returns A promise that resolves with the response.
  */
 export async function shopifyFetch(url: RequestInfo, init?: RequestInit): Response {
   const options: RequestInit = {
     ...(init ?? {}),
     headers: {
-      ...buildHeaders(),
+      ...(await buildHeaders()),
       ...(init?.headers ?? {}),
     },
   }
